code/requirements.txt
pandas==2.2.2
faiss-cpu==1.8.0
sentence-transformers==2.7.0
requests==2.31.0
tqdm==4.66.4
numpy==1.26.4
python-dotenv==1.0.1
